refactor(web-app): add support for data-less requests in the useCosMutationRequest hook
